Create LICENSE.txt
Added a LICENSE.txt (MIT). That's the "default" per https://wiki.jenkins-ci.org/display/JENKINS/Before+starting+a+new+plugin but it is better (legally) to have an explicit license than relying on something written on some web page on the internet. License are for lawyers and lawyers prefer the explicit way, sadly.